Update Int+Hex.swift
Fix UInt16 conversion error.

+++ nRFMeshProvision/Classes/Type Extensions/Int+Hex.swift
@@ -73,7 +73,7 @@ internal extension UInt16 {
     
     var data: Data {
         var int = self
-        return Data(bytes: &int, count: MemoryLayout<UInt8>.size)
+        return Data(bytes: &int, count: MemoryLayout<UInt16>.size)
     }
     
 }
